ports/tm4c123: changed pin_mask to 32 bit

+++ ports/tm4c123/mods/pin.h
@@ -97,7 +97,7 @@ typedef struct {
     uint32_t            dir;
     uint32_t            type;
     uint32_t            drive;
-    const uint8_t       pin_mask;
+    const uint32_t      pin_mask;
     const uint8_t       pin_num;
     int8_t              af;
     const uint8_t       num_afs;     // 255 AFs
